+++ src/app/nav/navigation.component.ts
@@ -1,6 +1,8 @@
 import { Component } from '@angular/core';
 
+import { APP_NAME } from '../app.constants';
 import { NAVIGATION_MENU } from './nav';
+import { getVersion } from '../shared/helpers/version';
 
 @Component({
     selector: 'ms-navigation',
@@ -9,6 +11,8 @@ import { NAVIGATION_MENU } from './nav';
 })
 export class NavigationComponent {
 
+    appName = APP_NAME;
+    appVersion = getVersion();
     navigationMenu = NAVIGATION_MENU;
 
 }