+++ src/app/shared/helpers/version.tns.ts
@@ -1,0 +1,3 @@
+import { getVersionNameSync } from 'nativescript-appversion';
+
+export const getVersion = getVersionNameSync;