+++ src/environments/environment.prod.ts
@@ -1,4 +1,5 @@
 export const environment = {
     production: true,
     backendUrl: '',
+    version: '1.0.0',
 };