+++ src/environments/environment.ts
@@ -3,4 +3,5 @@ declare const PORT;
 export const environment = {
     production: false,
     backendUrl: `http://localhost:${PORT}`,
+    version: '1.0.0-dev',
 };